Xxxb ranch

+++ Entities/Concrete/Besinler.cs
@@ -12,11 +12,11 @@ namespace Entities.Concrete
     {
         public int ID { get;set; }
         public string BesinAdi { get;set; }
-        public float GramKarsiligi { get;set; }
-        public float Kalori { get;set; }
-        public float Karbonhidrat { get;set; }
-        public float Protein { get;set; }
-        public float Yag { get;set; }
+        public decimal GramKarsiligi { get;set; }
+        public decimal Kalori { get;set; }
+        public decimal Karbonhidrat { get;set; }
+        public decimal Protein { get;set; }
+        public decimal Yag { get;set; }
         public List<TuketilenBesinler> TuketilenBesinler { get;set; }
         public DateTime CreatedDate { get;set; }=DateTime.Now;
         public DateTime? ModifiedDate { get;set; }
